change parameter to nonce buffer
